src/requirements.txt
# Crypto Trader Bot — Dependencies
# Python 3.11+

# Bybit API
pybit==5.8.0

# Telegram уведомления
python-telegram-bot==21.5

# Конфигурация
python-dotenv==1.0.1

# Планировщик задач (ежедневная сводка)
APScheduler==3.10.4

# Асинхронность
aiohttp==3.10.5
websockets==12.0

# Утилиты
colorama==0.4.6
